Added cell efp tooltip
