+++ Eplant/UI/Layout/ViewContainer/index.tsx
@@ -56,7 +56,6 @@ export function ViewContainer<T, S, A>({
 } & BoxProps) {
   const { activeData, error, loading, loadingAmount, dispatch, state } =
     useViewData(view, gene)
-
   const idLabel = React.useId()
   const selectId = React.useId()
   const [printing, setPrinting] = usePrinting()
@@ -262,7 +261,7 @@ export function ViewContainer<T, S, A>({
               <view.component
                 state={state}
                 geneticElement={gene}
-                activeData={activeData}
+                activeData={{...activeData, efps: view.efps}}
                 dispatch={dispatch}
               />
             </>
